final commit - added swagger ui docs

+++ src/app/useCases/User/deleteUser/DeleteUserUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from "tsyringe";
 
 import { IUsersRepository } from "../../../../database/repositories/IUserRepository";
+import { AppError } from "../../../../shared/errors/AppError";
 
 @injectable()
 class DeleteUserUseCase {
@@ -9,6 +10,12 @@ class DeleteUserUseCase {
     private usersRepository: IUsersRepository
   ) {}
   async execute(id: string): Promise<void> {
+    const userExists = await this.usersRepository.findById(id);
+
+    if (!userExists) {
+      throw new AppError("User not found", 404);
+    }
+
     await this.usersRepository.deleteUser(id);
   }
 }
