+++ src/app/useCases/User/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -18,7 +18,7 @@ class UpdateUserAvatarUseCase {
   async execute({ user_id, avatar_file }: IRequest): Promise<void> {
     const user = await this.usersRepository.findById(user_id);
 
-    if (user.avatar) {
+    if (user.avatar && user.avatar != null) {
       await deleteFile(`./tmp/avatar/${user.avatar}`);
     }
